Simpler way to add Up Button in a lower-level activity
Following:
https://developer.android.com/training/basics/actionbar/adding-buttons.html#UpNav
Instead of StatusActivity extending SubActivity, just extend Activity,
then directly enable Up Button for the Action Bar in the onCreate()
callback. Finally, declare in AndroidManifest.xml the parent activity,
for the StatusActivity, which would be MainActivity.

So we could avoid needing SubActivity, and avoid needing to implement
its onOptionsItemSelected() callback to go to MainActivity when
Up-Home is pressed. (Note we've not refactored the other lower-level
subactivity classes so they all still extend SubActivity.)